refactor(slideshow): organize StepExperience components and add reusable navigation buttons
- Moved all StepExperience-related components into a dedicated folder for better structure and maintainability.
- Created a reusable NavigationButton component for "Prev" and "Next" actions.
- Replaced individual navigation buttons in all Steps components with the new reusable component.
