+++ client/src/features/slideshow-form/components/StepNavigationButtons.tsx
@@ -1,4 +1,4 @@
-import { Box } from "@chakra-ui/react";
+import { Box, Flex } from "@chakra-ui/react";
 import BaseButton from "@/components/ui/BaseButton";
 
 interface StepNavigationButtonsProps {
@@ -14,35 +14,28 @@ interface StepNavigationButtonsProps {
 const StepNavigationButtons = ({
   onPrevStep,
   onNextStep,
-  prevButtonText = "Back",
-  nextButtonText = "Next",
-  isLoading = false,
-  isPrevDisabled = false,
-  isNextDisabled = false,
+  
 }: StepNavigationButtonsProps) => {
   return (
-    <Box display="flex" justifyContent="space-between" pt={6}>
-      <BaseButton
-        onClick={onPrevStep}
-        colorScheme="red"
-        variant="outline"
-        size="lg"
-        minW="140px"
-        // disabled={isPrevDisabled}
-      >
-        {prevButtonText}
-      </BaseButton>
-      <BaseButton
-        onClick={onNextStep}
-        size="lg"
-        variant="subtle"
-        colorScheme="green"
-        minW="120px"
-        // disabled={isNextDisabled || isLoading}
-      >
-        {nextButtonText}
-      </BaseButton>
-    </Box>
+    <Flex justify="space-between" pt={8} gap={4}>
+    <BaseButton
+      onClick={onPrevStep}
+      colorPalette="red"
+      variant="outline"
+      size="lg"
+      minW="140px"
+    >
+      Back
+    </BaseButton>
+    <BaseButton
+      onClick={onNextStep}
+      size="lg"
+      colorPalette="blue"
+      minW="140px"
+    >
+      Next
+    </BaseButton>
+  </Flex>
   );
 };
